new version

+++ src/components/transactions/TransactForm.js
@@ -159,7 +159,7 @@ export const TransactForm = () => {
                             store: currentTransact.storeId,
                             transaction_type: currentTransact.transaction_typeId,
                             transaction_date: currentTransact.transaction_date,
-                            amount: parseInt(currentTransact.amount)
+                            amount: parseFloat(currentTransact.amount)
                         }
 
                         //Send POST request to your API
